created custom hook to get data
