+++ components/useDeleteData.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const useDeleteData = () => {
+
+}
+
+export default useDeleteData
